Fix div on decimal

+++ math.h
@@ -12,7 +12,7 @@ int mul(int a, int b){
 	return a * b;
 }
 
-int div(int a, int b){
-	return a / b;
+double div(int a, int b){
+	return (double)a / b;
 }
